Adding in comments to explain functions
