+++ detectree2/models/reproject.py
@@ -121,8 +121,7 @@ def reproject_to_geojson(directory = None, EPSG = "26917"):
 # Reprojects the coordinates back so the crowns can be overlaid with the original tif file of the entire region
 def reproject_to_geojson_spatially(data, output_fold = None, pred_fold = None, EPSG = "26917"):
     """
-    Takes a json and changes it to a geojson so it can overlay with crowns
-    Another copy is produced to overlay with PNGs
+    Takes a json and changes it to a geojson so it can overlay with crowns of the original tif
     """
 
     Path(output_fold).mkdir(parents=True, exist_ok=True)
@@ -208,7 +207,7 @@ def reproject_to_geojson_spatially(data, output_fold = None, pred_fold = None, E
             # Check final form is correct - compare to a known geojson file if error appears.
             # print("geofile",geofile)
 
-            output_geo_file = output_fold + img_dict["filename"].replace('.json',"_"+EPSG+'.geojson')
+            output_geo_file = output_fold + img_dict["filename"].replace('.json',"_"+EPSG+'_lidar.geojson')
             # print("output location:", output_geo_file)
             with open(output_geo_file, "w") as dest:
               json.dump(geofile,dest)
